move initViewAs back to progress.js for ease of testing
